refactor: To improve readability, split into header and source files
